% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMLE.R
\name{reptable}
\alias{reptable}
\title{Replicate the table 1 and 2}
\usage{
reptable(gamma = 0, n = 200, rep = 8, clusterN = 8)
}
\arguments{
\item{gamma}{the tradeoff parameter}

\item{n}{the sample size}

\item{rep}{the replication number}

\item{clusterN}{the number of cluster}
}
\value{
The estimator, SE, SEE, and the  coverage probability of the $95\%$ confidence interval based on the asymtotic normal approximation.
}
\description{
Replicate the table 1 and 2
}
\examples{
# Reproduce table 1
set.seed(47)
for(gamma in seq(0,1,length.out = 5)){
  for(n in c(200,400)){
    reptable(gamma = gamma, n = n,rep = 1000,clusterN = 100 )
  }
}
# Reproduce table 2
for(gamma in seq(0,1,length.out = 3)){
  for(n in c(400)){
    reptable(gamma = gamma, n = n,rep = 1000,clusterN = 100 )
  }
}
}
